# Use atomic operations in NewBasic, remove overflow check and add benc…