…hmarks.

Fixes #21.

   u@x1 ~/g/s/e/ecs> benchcmp old.txt new.txt
   benchmark               old ns/op     new ns/op     delta
   BenchmarkNewBasic-4     27.3          10.4          -61.90%

Also add benchmarks to find the breakpoint where it starts to make
sense to allocate several basic entities. On this machine, using
NewEntities is faster even for 10 entities as it only uses a single
lock as compared to 10 atomic operations.

BenchmarkNewBasic-4        	100000000	        10.5 ns/op
BenchmarkNewBasics1-4      	30000000	        42.8 ns/op
BenchmarkNewBasic10-4      	20000000	       105 ns/op
BenchmarkNewBasics10-4     	20000000	        70.3 ns/op
BenchmarkNewBasic100-4     	 1000000	      1046 ns/op
BenchmarkNewBasics100-4    	 3000000	       433 ns/op
BenchmarkNewBasic1000-4    	  200000	     10418 ns/op
BenchmarkNewBasics1000-4   	  300000	      3907 ns/op