@mewmew 2016-05-13
Reference discussion about removing the overflow check https://gitter.im/EngoEngine/engo?at=57363c2a1794136a7d087d6a

> When you add entities at full clock-speed, engo will run for 10 years without overflow
>
> (assuming adding an entity takes only 1nanosecond, in practice it takes longer)
>
> Oh, did I say 10 years?
>
> Sorry, meant 100 :worried: